@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 
-import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -25,22 +24,23 @@ import java.util.stream.Stream;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
- class User {
+class User {
 
     private Integer id;
     private String name;
     private String pwd;
 
-    private Boolean excellent;
+    private Boolean isVIP;
 
 }
+
 public class StreamAPI {
 
     @SneakyThrows
-    public static void main(String[] args)  {
+    public static void main(String[] args) {
         ArrayList<User> users = new ArrayList<>();
-        users.add(new User(1, "a", "111",true));
-        users.add(new User(2, "b", "222",true));
+        users.add(new User(1, "a", "111", true));
+        users.add(new User(2, "b", "222", true));
         users.add(new User(3, "c", "333", false));
 
         /**
@@ -49,19 +49,18 @@ public class StreamAPI {
         List<String> collect = users.stream()
                 .filter(it -> it.getId() < 10)
                 .sorted(Comparator.comparing(User::getId))  //从小到大排序
-                .sorted(new Comparator<User>() {
-                    @Override
-                    public int compare(User o1, User o2) {
-                        return o1.getId() - o2.getId();
-                    }
-                })
-                .map(User::getName)  //将每一个元素映射为 name,代替原来的元素
-                .collect(Collectors.toList());
+                .sorted(Comparator.comparing(User::getName))
+                .map(User::getName)
+                .toList();
+        Map<Boolean, List<User>> collect1 = users.stream().collect(Collectors.partitioningBy(user -> user.getIsVIP()));
+        //根据是不是vip划分
 
+        Map<Integer, List<User>> collect2 = users.stream().collect(Collectors.groupingBy(user -> user.getId()));
+        //根据id 分为多组
 
-        collect.forEach(System.out::println);
+        //  collect.forEach(System.out::println);
         Map<Boolean, List<User>> map = users.stream()
-                .collect(Collectors.groupingBy(User::getExcellent));
+                .collect(Collectors.groupingBy(User::getIsVIP));
         //将list按照属性分组 ,返回一个map ,key是属性,value是对应的相同属性的元素集合
 
         map.forEach((k, v) -> System.out.println(k + ":" + v));  //遍历map
@@ -69,15 +68,22 @@ public class StreamAPI {
         /**
          * 除list生成流,还有其他方式
          */
-        Stream<int[]> stream =  Stream.of(new int[10]);//从数组中获取stream
+        Stream<int[]> stream = Stream.of(new int[10]);//从数组中获取stream
         Stream<Integer> integerStream = Stream.of(1, 2, 3);
+        //泛型表示每个元素的类型   of一个基本类型数组 只有一个元素
+        Stream<String> stringStream = Stream.of(new String[10]);
+        Stream<char[]> stream1 = Stream.of("xxx".toCharArray());
 
-        try {
-            Stream<String> lines = Files.lines(Paths.get("test.txt"), Charset.defaultCharset());//从文件中获取stream
-        } catch (IOException e) {
-            System.out.println("none");
-        }
 
+        //文件内容生成
+        Stream<String> lines = Files.lines(
+                Paths.get("src/test.txt"),
+                Charset.defaultCharset()
+        );//从文件中获取stream
+        lines.forEach(System.out::println);
+        /*
+        无限流   初始值迭代  / 函数生成
+         */
         Stream<Integer> limit = Stream.iterate(0, n -> n + 2).limit(5);
         limit.forEach(System.out::println); //通过初始值迭代生成
 
@@ -101,7 +107,10 @@ public class StreamAPI {
          * 通过min/max获取最小最大值
          * foreach进行元素遍历
          */
-
+//        Files.lines(
+//                Paths.get("C:\\Users\\ouyan\\IdeaProjects\\demo1\\bootdemo\\src\\test.txt"),
+//                Charset.defaultCharset()
+//        ).forEach(System.out::println);
 
     }
 }
