@@ -4,12 +4,14 @@ import com.ouyang.boot.entity.Account;
 import com.ouyang.boot.entity.AccountRepository;
 import com.ouyang.boot.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.web.server.Cookie;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap ;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 2022-02-11 20:54
@@ -27,18 +29,17 @@ public class MyRestController {
     public HashMap<String, Object> pic(@PathVariable String id,
                                        @PathVariable String name,
                                        @RequestHeader("User-Agent") String userAgent,
-                                       @RequestHeader HashMap<String, String> header
-                                       //@RequestAttribute("aaa") Object aaa,// 获取请求域中的属性,用于同一请求下不同页面的数据交换
-                                       // @CookieValue Cookie cookie
-    ) {
-        HashMap<String, Object> map = new HashMap<>(10);
-        map.put("id", id);
-        map.put("name", name);
-        map.put("userAgent", userAgent);
-        map.put("header", header);
-        // map.put("cookie", cookie);
+                                       @RequestHeader HashMap<String, String> header,
+                                       @RequestAttribute("aaa") Object aaa,
+                                       // 获取请求域中的属性,用于同一请求下不同页面的数据交换
+                                       @CookieValue Cookie cookie) {
 
-        return map;
+            return new HashMap<String,Object>(){{
+                put("id", id);
+                put("name", name);
+                put("userAgent", userAgent);
+                put("header", header);
+            }};
     }
 
     @Autowired
